extract loader-v3-interface crate (#4484)
* extract loader-v3-interface

* fix doctests

* update path in doc comment


---------
