+++ program/Cargo.toml
@@ -54,6 +54,7 @@ solana-instruction = { workspace = true, default-features = false, features = [
 solana-instructions-sysvar = { workspace = true }
 solana-keccak-hasher = { workspace = true, features = ["sha3"] }
 solana-last-restart-slot = { workspace = true, features = ["serde", "sysvar"] }
+solana-loader-v3-interface = { workspace = true, features = ["bincode"] }
 solana-message = { workspace = true, features = ["bincode", "blake3"] }
 solana-msg = { workspace = true }
 solana-native-token = { workspace = true }
